+++ CHANGELOG.md
@@ -30,6 +30,9 @@
 ## v0.152.0
 - Updates OpenTelemetry modules to [v1.58.0/v0.152.0](https://github.com/open-telemetry/opentelemetry-collector/releases/tag/v0.152.0)
 
+### Breaking Changes
+- ⚠️ **swok8sdiscovery receiver** : The `DiscoveredBy` relation now has inverted direction.
+
 ## v0.150.4
 - **Added** `swootelentityrefprocessor` for manipulating EntityRefs